+++ changelog.md
@@ -24,6 +24,7 @@
 - Various project metadata
 - Full CI pipeline
 - License checker in CI
+- BackendDispatcher object to help with pandas / koalas manipulation
 
 ### Fixed